refactor: rename to state-factory
