+++ src/data/loom-file.ts
@@ -1,6 +1,6 @@
 import { App, Notice, normalizePath } from "obsidian";
 import { createFile, createFolder } from "./file-operations";
-import { createLoomState } from "../shared/loom-state/loom-state-factory";
+import { createLoomState } from "../shared/loom-state/state-factory";
 import { serializeState } from "./serialization";
 import { LOOM_EXTENSION, DEFAULT_LOOM_NAME } from "./constants";
