+++ src/shared/loom-state/commands/column-add-command/column-add-command.spec.ts
@@ -1,4 +1,4 @@
-import { createTestLoomState } from "src/shared/loom-state/loom-state-factory";
+import { createTestLoomState } from "src/shared/loom-state/state-factory";
 import CommandUndoError from "../command-undo-error";
 import CommandRedoError from "../command-redo-error";
 import ColumnAddCommand from ".";
