+++ src/shared/loom-state/commands/column-type-update-command.spec.ts
@@ -2,7 +2,7 @@ import {
 	createTestLoomState,
 	createTag,
 	createTextFilter,
-} from "src/shared/loom-state/loom-state-factory";
+} from "src/shared/loom-state/state-factory";
 import CommandUndoError from "./command-undo-error";
 import CommandRedoError from "./command-redo-error";
 import ColumnTypeUpdateCommand from "./column-type-update-command";
