+++ src/shared/loom-state/commands/row-add-command.spec.ts
@@ -1,4 +1,4 @@
-import { createTestLoomState } from "src/shared/loom-state/loom-state-factory";
+import { createTestLoomState } from "src/shared/loom-state/state-factory";
 import RowAddCommand from "./row-add-command";
 import CommandUndoError from "./command-undo-error";
 import CommandRedoError from "./command-redo-error";
