+++ src/shared/loom-state/commands/row-insert-command.ts
@@ -1,7 +1,4 @@
-import {
-	createCell,
-	createRow,
-} from "src/shared/loom-state/loom-state-factory";
+import { createCell, createRow } from "src/shared/loom-state/state-factory";
 import LoomStateCommand from "./loom-state-command";
 import { Row, LoomState, SortDir } from "../types/loom-state";
 import RowNotFoundError from "src/shared/error/row-not-found-error";
