+++ src/shared/loom-state/commands/tag-cell-multiple-remove-command.spec.ts
@@ -1,7 +1,7 @@
 import {
 	createTestLoomState,
 	createTag,
-} from "src/shared/loom-state/loom-state-factory";
+} from "src/shared/loom-state/state-factory";
 import CommandUndoError from "./command-undo-error";
 import { advanceBy, clear } from "jest-date-mock";
 import TagCellMultipleRemoveCommand from "./tag-cell-multiple-remove-command";
