+++ src/shared/loom-state/state-factory.ts
